requirements.txt
psycopg2-binary==2.9.9
aiogram==3.2.0
requests==2.31.0
emoji==2.9.0